dev_requirements.txt
apache-airflow[postgres]
pendulum
requests==2.28.2
python-dotenv==0.21.1
psycopg2-binary==2.9.5
pydantic==1.10.5